+++ src/V1/ResourceIdentifier.php
@@ -47,20 +47,20 @@ final class ResourceIdentifier extends AbstractElement
             throw new ValidationException('A resource object MUST contain a type');
         }
 
+        if (! is_string($object->type)) {
+            throw new ValidationException('A resource type MUST be a string');
+        }
+
         if (! property_exists($object, 'id')) {
             throw new ValidationException('A resource object MUST contain an id');
         }
 
-        if (is_object($object->type) or is_array($object->type)) {
-            throw new ValidationException('Resource type cannot be an array or object');
+        if (! is_string($object->id)) {
+            throw new ValidationException('A resource id MUST be a string');
         }
 
-        if (is_object($object->id) or is_array($object->id)) {
-            throw new ValidationException('Resource Id cannot be an array or object');
-        }
-
-        $this->set('type', strval($object->type));
-        $this->set('id', strval($object->id));
+        $this->set('type', $object->type);
+        $this->set('id', $object->id);
 
         if (property_exists($object, 'meta')) {
             $this->set('meta', $this->create('Meta', $object->meta));