+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING**: Providing the fields `type` or `id` in a resource not as a string throws a `\Art4\JsonApiClient\Exception\ValidationException`, provide them always as strings instead
+
 ### Removed
 
 - **BREAKING**: `Art4\JsonApiClient\AccessInterface` removed, use `Art4\JsonApiClient\Accessable` instead
@@ -60,6 +64,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **BREAKING**: `Art4\JsonApiClient\Utils\Helper` removed, use `Art4\JsonApiClient\Helper\Parser` instead
 - **BREAKING**: `Art4\JsonApiClient\Utils\Manager` removed, use `Art4\JsonApiClient\Manager\ErrorAbortManager` instead
 - **BREAKING**: `Art4\JsonApiClient\Utils\ManagerInterface` removed, use `Art4\JsonApiClient\Manager` instead
+
+## [0.10.1] - 2019-09-23
+
+### Deprecated
+
+- Providing the fields `type` or `id` in a resource not as a string will be throw a ValidationException in v1.0, provide them always as strings instead
 
 ## [0.10] - 2018-11-07
 
@@ -304,7 +314,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Validator fits nearly 100% specification
 - Full test coverage
 
-[Unreleased]: https://github.com/Art4/json-api-client/compare/0.10...HEAD
+[Unreleased]: https://github.com/Art4/json-api-client/compare/0.10.1...HEAD
+[0.10.1]: https://github.com/Art4/json-api-client/compare/0.10...0.10.1
 [0.10]: https://github.com/Art4/json-api-client/compare/0.9.1...0.10
 [0.9.1]: https://github.com/Art4/json-api-client/compare/0.9...0.9.1
 [0.9]: https://github.com/Art4/json-api-client/compare/0.8.1...0.9